+++ main.js
@@ -11,9 +11,10 @@ define(function (require, exports, module) {
     var CommandManager = brackets.getModule("command/CommandManager"),
 		DocumentManager = brackets.getModule("document/DocumentManager"),
 		Editor = brackets.getModule("editor/EditorManager"),
-        Menus          = brackets.getModule("command/Menus"),
-        FileSystem                 = brackets.getModule("filesystem/FileSystem"),
-        FileUtils                  = brackets.getModule("file/FileUtils"),
+        Menus = brackets.getModule("command/Menus"),
+        FileSystem = brackets.getModule("filesystem/FileSystem"),
+        FileUtils = brackets.getModule("file/FileUtils"),
+        StatusBar = brackets.getModule("widgets/StatusBar"),
         KeyBindingManager  = brackets.getModule("command/KeyBindingManager"),
 		AppInit        = brackets.getModule("utils/AppInit"),
 		PanelManager = brackets.getModule("view/PanelManager"),
@@ -39,6 +40,8 @@ define(function (require, exports, module) {
     prefs.definePreference("debugger-host", "string", "localhost");
     prefs.definePreference("showChangelogOnUpdate", "boolean", true);
     prefs.definePreference("lastVersion", "string", "none");
+    prefs.definePreference("autoConnectOnToggle", "boolean", false);
+    prefs.definePreference("autoConnect", "boolean", false);
 	
 	var nodeDebuggerDomain = new NodeDomain("brackets-node-debugger", ExtensionUtils.getModulePath(module, "node/main"));
 	
@@ -61,20 +64,37 @@ define(function (require, exports, module) {
 		breakpointGutters.init(nodeDebuggerDomain);
         nodeDebuggerPanel.init(nodeDebuggerDomain, $logPanel);
 
+        //AutoConnect
+        if(prefs.get("autoConnect")) {
+            nodeDebuggerDomain.exec("start", prefs.get("debugger-port"), prefs.get("debugger-host"), true);
+        var $sb = $("<div>").addClass("ion-android-developer");
+        StatusBar.addIndicator("node-debugger-indicator", $sb, true, null, "Node.js Debugger");
+        }
+
 		$(nodeDebuggerDomain).on("connect", function() {
 			nodeDebuggerPanel.log( $('<span>').text('Debugger connected') );
             $logPanel.find('.activate').addClass('ion-ios7-checkmark')
                                     .removeClass('ion-ios7-close');
             $logPanel.find('a.inactive').addClass('active').removeClass('inactive');
+            $('#node-debugger-indicator').addClass('connected');
 		});
 
-		$(nodeDebuggerDomain).on("close", function() {
+		$(nodeDebuggerDomain).on("close", function(e, err) {
+            var msg = "Debugger disconnected";
+            if(err) {
+                msg += ": " + err;
+            }
+
+            if(err === 'ECONNREFUSED') {
+                msg = "Couldn't connect to " + prefs.get("debugger-port") + ":" + prefs.get("debugger-host");
+            }
             breakpointGutters.removeAllBreakpoints();
-			nodeDebuggerPanel.log( $('<span>').text('Debugger disconnected') );
+			nodeDebuggerPanel.log( $('<span>').text(msg) );
 
             $logPanel.find('.activate').addClass('ion-ios7-close')
                                     .removeClass('ion-ios7-checkmark');
             $logPanel.find('a.active').addClass('inactive').removeClass('active');
+            $('#node-debugger-indicator').removeClass('connected');
 
             //remove highlight
             if(highlightCm) {
@@ -148,7 +168,7 @@ define(function (require, exports, module) {
 
 		$logPanel.find('.activate').on('click', function() {
 			//Starts the socket and connects to the V8 debugger
-			nodeDebuggerDomain.exec("start", prefs.get("debugger-port"), prefs.get("debugger-host"));
+			nodeDebuggerDomain.exec("start", prefs.get("debugger-port"), prefs.get("debugger-host"), false);
 		});
 		
 		$logPanel.find('.next').on('click', function() {
@@ -170,6 +190,11 @@ define(function (require, exports, module) {
 			nodeDebuggerDomain.exec('continue');
             debuggerContinue();
 		});
+
+        //Open panel on status indicator click
+        $('#node-debugger-indicator').on('click', function() {
+            toggleLog();
+        });
 	});
 
     // Function to run when the menu item is clicked